+++ mvc/BaseCodedException.php
@@ -12,6 +12,7 @@ namespace sinri\enoch\mvc;
 class BaseCodedException extends \Exception
 {
     const DEFAULT_ERROR = 0;
+    const NOT_IMPLEMENT_ERROR = 1;
 
     const USER_NOT_LOGIN = 101;
     const USER_NOT_ADMIN = 102;