+++ mvc/Naamah.php
@@ -8,16 +8,15 @@
 
 namespace sinri\enoch\mvc;
 
-use sinri\enoch\core\Spirit;
-
 /**
  * Class Naamah
  * Lamech's daughter with Zillah
  * REGEX FOR PATH would not add ^$ to two sides.
  * Route Manager
+ * @deprecated sinri v1.2.0
  * @package sinri\enoch\mvc
  */
-class Naamah
+class Naamah extends RouterInterface
 {
     const ROUTE_PARAM_TYPE = "type";
     const ROUTE_PARAM_REGEX = "regex";
@@ -28,25 +27,9 @@ class Naamah
     const ROUTE_TYPE_FUNCTION = 'function';
     const ROUTE_TYPE_VIEW = 'view';
 
-    private $error_handler = null;
-
-    protected $spirit = null;
-
-    /**
-     * @param null $errorHandler
-     */
-    public function setErrorHandler($errorHandler)
-    {
-        $this->error_handler = $errorHandler;
-    }
-
-    protected $routes = [];
-
     public function __construct()
     {
-        $this->routes = [];
-        $this->error_handler = null;
-        $this->spirit = new Spirit();
+        parent::__construct();
     }
 
     /**
@@ -112,15 +95,5 @@ class Naamah
         throw new BaseCodedException("No route matched.", BaseCodedException::NO_MATCHED_ROUTE);
     }
 
-    public function handleRouteError($errorData = [])
-    {
-        if (is_string($this->error_handler) && file_exists($this->error_handler)) {
-            $this->spirit->displayPage($this->error_handler, $errorData);
-            return;
-        } elseif (is_callable($this->error_handler)) {
-            call_user_func_array($this->error_handler, [$errorData]);
-            return;
-        }
-        $this->spirit->errorPage(__METHOD__);
-    }
+
 }