+++ test/routing/index.php
@@ -7,5 +7,6 @@
  */
 define("GATEWAY_NAME", "index.php");
 date_default_timezone_set("Asia/Shanghai");
-require_once __DIR__ . '/index_route.php';
+//require_once __DIR__ . '/index_route.php';
 //require_once __DIR__.'/index_ci.php';
+require_once __DIR__ . '/index_adah.php';